requirements.txt
# pip install -r "requirements.txt"

atlassian-python-api
beautifulsoup4
lxml
pypdf
requests
requests_ntlm
